pyproject.toml
[project]
name = "llmbotrewrite"
version = "0.1.0"
description = ""
requires-python = ">=3.13"
dependencies = [
    "discord.py",
    "aiohttp",
    "yarl",
    "ollama",
    "pillow",
    "pydantic",
]

[dependency-groups]
dev = [
    "ruff",
]

[tool.setuptools]
py-modules = [
    "bot",
    "functionality",
]

[tool.ruff]
target-version = "py313"

line-length = 120
indent-width = 4

